+++ CubA4Loader/src/window/Window.hpp
@@ -13,14 +13,14 @@ namespace CubA4
 			public virtual IWindow
 		{
 		public:
-			static std::shared_ptr<Window> createWindow(int sizeX, int sizeY, int64_t additionalSDLWindowFlags);
+			static std::shared_ptr<Window> createWindow(int sizeX, int sizeY, uint32_t additionalSDLWindowFlags);
 
 			SDL_Window *getSDLWindow();
 			bool isFullscreen();
 			void setFullscreen(bool fullscreen);
 			void toggleFullscreen();
 		private:
-			explicit Window(int sizeX, int sizeY, int64_t additionalSDLWindowFlags);
+			explicit Window(int sizeX, int sizeY, uint32_t additionalSDLWindowFlags);
 			~Window();
 
 			bool fullscreenMode_;
